refactor: Calendar 중복되는 style 코드 공통화
