+++ frontend/src/components/CoachCalendar/index.tsx
@@ -1,6 +1,7 @@
 import { useState, useMemo, useEffect } from 'react';
 
 import * as S from './styled';
+import * as C from '../@common/CalendarStyle/styled';
 
 import {
   useCalendarActions,
@@ -61,17 +62,17 @@ const CoachCalendar = () => {
 
   return (
     <S.Box>
-      <S.Header>
-        <S.MonthPicker onClick={handleClickMonthPicker}>{monthNames[month]}</S.MonthPicker>
-        <S.YearPicker>
-          <S.YearChange onClick={handleClickPrevYear}>{'<'}</S.YearChange>
+      <C.Header>
+        <C.MonthPicker onClick={handleClickMonthPicker}>{monthNames[month]}</C.MonthPicker>
+        <C.YearPicker>
+          <C.YearChange onClick={handleClickPrevYear}>{'<'}</C.YearChange>
           <p>{year}</p>
-          <S.YearChange onClick={handleClickNextYear}>{'>'}</S.YearChange>
-        </S.YearPicker>
-      </S.Header>
+          <C.YearChange onClick={handleClickNextYear}>{'>'}</C.YearChange>
+        </C.YearPicker>
+      </C.Header>
 
-      <S.Body>
-        <S.WeekDay>
+      <C.Body>
+        <C.WeekDay>
           <div>Sun</div>
           <div>Mon</div>
           <div>Tue</div>
@@ -79,8 +80,8 @@ const CoachCalendar = () => {
           <div>Thu</div>
           <div>Fri</div>
           <div>Sat</div>
-        </S.WeekDay>
-        <S.Days key={rerenderKey}>
+        </C.WeekDay>
+        <C.Days key={rerenderKey}>
           {Array.from({ length: daysLength }, (_, index) => {
             if (isOverFirstDay(index)) {
               const day = getDay(index);
@@ -94,46 +95,46 @@ const CoachCalendar = () => {
 
               if (isToday(day)) {
                 return (
-                  <S.Day key={index} today>
+                  <S.CalendarDay key={index} today>
                     {day}
                     {reservations}
-                  </S.Day>
+                  </S.CalendarDay>
                 );
               }
 
               if (isBeforeToday(day)) {
                 return (
-                  <S.Day key={index} type="disable">
+                  <S.CalendarDay key={index} type="disable">
                     {day}
                     {reservations}
-                  </S.Day>
+                  </S.CalendarDay>
                 );
               }
 
               return (
-                <S.Day key={index}>
+                <S.CalendarDay key={index}>
                   {day}
                   {reservations}
                   <span />
                   <span />
                   <span />
                   <span />
-                </S.Day>
+                </S.CalendarDay>
               );
             }
 
-            return <S.Day key={index} />;
+            return <S.CalendarDay key={index} />;
           })}
-        </S.Days>
-      </S.Body>
+        </C.Days>
+      </C.Body>
 
-      <S.MonthContainer show={showMonthPicker}>
+      <C.MonthContainer show={showMonthPicker}>
         {monthNames.map((monthName, index) => (
           <div key={index} onClick={getHandleClickMonth(index)}>
             {monthName}
           </div>
         ))}
-      </S.MonthContainer>
+      </C.MonthContainer>
     </S.Box>
   );
 };
