+++ frontend/src/types/domain.ts
@@ -41,3 +41,5 @@ export interface CoachScheduleRequestBodyType {
 export type StringDictionary = {
   [key: string]: string[];
 };
+
+export type DayType = 'default' | 'disable' | 'active';
